+++ lesson13_MigratingDB/FabricMarket_DomainModel/Models/FabricProduction/FabricVariant.cs
@@ -1,6 +1,6 @@
 ﻿namespace lesson11_FabricMarket_DomainModel.Models.FabricProduction
 {
-    public class FabricVariant : Entity<long>
+    public class FabricVariant : Entity<int>
     {
         public required string Name { get; set; }
         public required string Description { get; set; }
