+++ lesson13_MigratingDB/FabricMarket_DomainModel/Models/FabricProduction/Vendor.cs
@@ -1,6 +1,6 @@
 ﻿namespace lesson11_FabricMarket_DomainModel.Models.FabricProduction
 {
-    public class Vendor : Entity<long>
+    public class Vendor : Entity<int>
     {
         public required string Name { get; set; }
